refactor: remove file-based real names connector and update nickname reveal logic
