+++ nicknamer/src/nicknamer/commands/names/mod.rs
@@ -35,3 +35,32 @@ impl NamesRepository for EmbeddedNamesRepository {
         Ok(names)
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use std::collections::HashMap;
+
+    #[test]
+    fn test_real_name_deser() {
+        // Create a YAML string representing RealNames
+        let yaml_data = r#"
+names:
+  123456789: Alice
+  987654321: Bob
+"#;
+
+        // Deserialize the YAML string using from_yaml
+        let deserialized: Names = serde_yml::from_str(yaml_data).unwrap();
+
+        // Create the expected RealNames object for comparison
+        let mut expected = Names {
+            names: HashMap::new(),
+        };
+        expected.names.insert(123456789, "Alice".to_string());
+        expected.names.insert(987654321, "Bob".to_string());
+
+        // Assert that deserialization produced the expected object
+        assert_eq!(deserialized, expected);
+    }
+}
