--- nicknamer/src/nicknamer/connectors/file/real_names.yml
@@ -1,21 +1,0 @@
-names:
-  102511651219058688: Matan
-  246711053898088448: Dušan
-  305511226841235456: Michael
-  363851243136745473: Alex
-  504405832071512064: Sid
-  778435459692625961: Andrew
-  894678255859273808: Clayton
-  894692357457469471: Zach
-  894771014859362326: Kyle
-  894942331973808181: Jake
-  899501665365929985: Kermit
-  779777332479000607: Erik
-  952694368152715355: Diana
-  1116357591140929628: Heather
-  1120392246659920074: Amy
-  1141570274270117889: Nicole
-  1141557841044447343: Jen
-  689624761134874642: Patton
-  1289942408875741214: MC
-  1291513296209444920: Joanna
