+++ nicknamer/src/nicknamer/connectors/mod.rs
@@ -1,2 +1,1 @@
 pub(crate) mod discord;
-pub(crate) mod file;
